pyproject.toml
[tool.poetry]
name = "async_sqs_consumer"
version = "0.3.6"
description = ""
authors = ["Diego Restrepo <drestrepo@fluidattacks.com>"]
repository = 'https://github.com/drestrepom/async_sqs_consumer'
readme = "README.md"
keywords = ["AWS", "AWS", "SQS", "sqs", "consumer", "async", "worker"]

[tool.poetry.dependencies]
python = "^3.9"
aiobotocore = "^2.4.0"
aioboto3 = "^10.1.0"
jsonschema = "^4.17.3"
pyrate-limiter = "^2.10.0"
aiohttp = "^3.8.4"
pytz = "^2023.3"

[tool.poetry.dev-dependencies]
prospector = "^1.7.7"
black = "^22.8.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
